pyproject.toml
[project]
name = "agente-vitalmed"
version = "0.1.0"
description = "Sistema multi-agentes para venda de planos de saúde e UTI móvel - Vitalmed"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # Framework de agentes
    "agno>=1.3.0",

    # API Gateway
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # Banco de dados
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.3.0",

    # Redis
    "redis>=5.0.0",

    # Google AI (Gemini + Embeddings)
    "google-generativeai>=0.8.0",
    "google-cloud-aiplatform>=1.60.0",

    # Observabilidade
    "langfuse>=2.0.0",

    # RAG - Parsers
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "openai-whisper>=20231117",

    # RAG - Embedding / Chunking
    "tiktoken>=0.7.0",

    # Integrações
    "httpx>=0.27.0",

    # Config / Utils
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",

    # Streamlit (testes/admin)
    "streamlit>=1.36.0",
    "pandas>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
